Updates after course

+++ src/main/java/fundamentals/TestSum.java
@@ -1,9 +1,10 @@
 package fundamentals;
 
 public class TestSum {
+    // Main method with big problems
     public static void main(String[] args) {
         // Initialize sum
-        float sum = 0;
+        double sum = 0;
         // Add 0.01, 0.02, ..., 0.99, 1 to sum
         for (var i = 0.01f; i <= 1.0d; i += 0.01d)
             sum += i;
@@ -11,4 +12,3 @@ public class TestSum {
         System.out.println("The sum is " + sum);
     }
 }
-
